Address UnnecessaryParentheses errorprone warnings
These prevent the roll of Chromium, see [1]. This CL is a follow-up of
https://webrtc-review.googlesource.com/c/src/+/363944 which didn't fix
all of the issues.

[1] - https://ci.chromium.org/ui/p/webrtc/builders/try/android_arm64_rel/78025/overview

Bug: None
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/363945
Auto-Submit: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#43101}

+++ sdk/android/instrumentationtests/src/org/webrtc/GlRectDrawerTest.java
@@ -51,8 +51,8 @@ public class GlRectDrawerTest {
       int width, int height, ByteBuffer actual, ByteBuffer expected) {
     actual.rewind();
     expected.rewind();
-    assertEquals(actual.remaining(), width * height * 3);
-    assertEquals(expected.remaining(), width * height * 3);
+    assertEquals(actual.remaining(), width * ((long) height) * 3L);
+    assertEquals(expected.remaining(), width * ((long) height) * 3L);
     for (int y = 0; y < height; ++y) {
       for (int x = 0; x < width; ++x) {
         final int actualR = actual.get() & 0xFF;
